requirements.txt
psycopg2-binary==2.9.5
requests==2.28.2
pandas==1.5.3
numpy==1.24.2
sqlalchemy==2.0.8
beautifulsoup4==4.12.2 
lxml==4.9.2
#pip install -r requirements.txt
